Seperate WeenyGui out from Weeny
