--- src/main/java/Weeny/ui/MainWindow.java
@@ -1,4 +1,0 @@
-package weeny.ui;
-
-public class MainWindow {
-}
